IAT-3083: ItemConstants have all been moved to ModelConstants in common
