+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/impl/ItemBankMapperImpl.java
@@ -15,9 +15,9 @@ import org.opentestsystem.ap.common.model.AssessmentItemCore;
 import org.opentestsystem.ap.common.model.FieldTestData;
 import org.opentestsystem.ap.common.model.Form;
 import org.opentestsystem.ap.common.model.Item;
-import org.opentestsystem.ap.common.model.ItemConstants;
 import org.opentestsystem.ap.common.model.ItemCore;
 import org.opentestsystem.ap.common.model.ItemMetadata;
+import org.opentestsystem.ap.common.model.ModelConstants;
 import org.opentestsystem.ap.common.model.ServerManaged;
 import org.opentestsystem.ap.common.model.StimItem;
 import org.opentestsystem.ap.common.model.TestAdministration;
@@ -62,7 +62,7 @@ public class ItemBankMapperImpl implements ItemBankMapper {
         final String author = itemUpdateEvent.getUserFullName().orElse(itemEntityResponse.getCreatedBy());
 
         if (existingItem == null) {
-            baseItem = ItemConstants.ItemType.TYPE_STIM.equals(item.getType()) ? new Stimulus() : new ImrtItem();
+            baseItem = ModelConstants.ItemType.TYPE_STIM.equals(item.getType()) ? new Stimulus() : new ImrtItem();
 
             baseItem.setId(convertItemId(item.getId()));
         } else {
