+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/keyword/EbsrKeywordContentMapper.java
@@ -26,7 +26,7 @@ import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.Ebsr
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.EbsrSection.SPANISH_PART_B_OPTIONS;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.EbsrSection.SPANISH_PART_B_STEM;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.EbsrSection.SPANISH_PROMPT;
-import static org.opentestsystem.ap.common.model.ItemConstants.ItemType.TYPE_EBSR;
+import static org.opentestsystem.ap.common.model.ModelConstants.ItemType.TYPE_EBSR;
 import static org.opentestsystem.ap.imrt.iis.core.keyword.KeywordContentUtil.convertContentToKeywordString;
 
 @Component
