+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/keyword/EqKeywordContentMapper.java
@@ -19,7 +19,7 @@ import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.EqSe
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.EqSection.SPANISH_LEFT_LABEL;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.EqSection.SPANISH_PROMPT;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.EqSection.SPANISH_RIGHT_LABEL;
-import static org.opentestsystem.ap.common.model.ItemConstants.ItemType.TYPE_EQ;
+import static org.opentestsystem.ap.common.model.ModelConstants.ItemType.TYPE_EQ;
 import static org.opentestsystem.ap.imrt.iis.core.keyword.KeywordContentUtil.convertContentToKeywordString;
 
 @Component
