+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/keyword/GiKeywordContentMapper.java
@@ -14,7 +14,7 @@ import java.util.Set;
 
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.GiSection.PROMPT;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.GiSection.SPANISH_PROMPT;
-import static org.opentestsystem.ap.common.model.ItemConstants.ItemType.TYPE_GI;
+import static org.opentestsystem.ap.common.model.ModelConstants.ItemType.TYPE_GI;
 import static org.opentestsystem.ap.imrt.iis.core.keyword.KeywordContentUtil.convertContentToKeywordString;
 
 @Component
