+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/keyword/HtqoKeywordContentMapper.java
@@ -20,7 +20,7 @@ import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.Htqo
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.HtqoSection.PROMPT;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.HtqoSection.SPANISH_INTERACTIVE_TEXTS;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.HtqoSection.SPANISH_PROMPT;
-import static org.opentestsystem.ap.common.model.ItemConstants.ItemType.TYPE_HTQO;
+import static org.opentestsystem.ap.common.model.ModelConstants.ItemType.TYPE_HTQO;
 import static org.opentestsystem.ap.imrt.iis.core.keyword.KeywordContentUtil.convertContentToKeywordString;
 
 @Component
