+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/keyword/HtqsKeywordContentMapper.java
@@ -19,7 +19,7 @@ import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.Htqs
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.HtqsSection.PROMPT;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.HtqsSection.SPANISH_INTERACTIVE_TEXTS;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.HtqsSection.SPANISH_PROMPT;
-import static org.opentestsystem.ap.common.model.ItemConstants.ItemType.TYPE_HTQS;
+import static org.opentestsystem.ap.common.model.ModelConstants.ItemType.TYPE_HTQS;
 import static org.opentestsystem.ap.imrt.iis.core.keyword.KeywordContentUtil.convertContentToKeywordString;
 
 @Component
