+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/keyword/McKeywordContentMapper.java
@@ -20,7 +20,7 @@ import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.McSe
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.McSection.PROMPT;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.McSection.SPANISH_OPTIONS;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.McSection.SPANISH_PROMPT;
-import static org.opentestsystem.ap.common.model.ItemConstants.ItemType.TYPE_MC;
+import static org.opentestsystem.ap.common.model.ModelConstants.ItemType.TYPE_MC;
 import static org.opentestsystem.ap.imrt.iis.core.keyword.KeywordContentUtil.convertContentToKeywordString;
 
 @Component
