+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/keyword/MiKeywordContentMapper.java
@@ -3,7 +3,6 @@ package org.opentestsystem.ap.imrt.iis.core.keyword;
 import org.opentestsystem.ap.common.imrt.model.BaseItem;
 import org.opentestsystem.ap.common.imrt.model.KeywordContent;
 import org.opentestsystem.ap.common.model.Item;
-import org.opentestsystem.ap.common.model.ItemConstants;
 import org.opentestsystem.ap.common.model.MiItem;
 import org.opentestsystem.ap.common.model.Table;
 import org.opentestsystem.ap.common.model.content.MiItemContent;
@@ -24,7 +23,7 @@ import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.MiSe
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.MiSection.SPANISH_TABLE_ROW_LABELS;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.MiSection.TABLE_COLUMN_HEADERS;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.MiSection.TABLE_ROW_LABELS;
-import static org.opentestsystem.ap.common.model.ItemConstants.ItemType.TYPE_MI;
+import static org.opentestsystem.ap.common.model.ModelConstants.ItemType.TYPE_MI;
 import static org.opentestsystem.ap.common.model.Table.CELL_TYPE_LABEL;
 import static org.opentestsystem.ap.imrt.iis.core.keyword.KeywordContentUtil.convertContentToKeywordString;
