+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/keyword/MsKeywordContentMapper.java
@@ -20,7 +20,7 @@ import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.MsSe
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.MsSection.PROMPT;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.MsSection.SPANISH_OPTIONS;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.MsSection.SPANISH_PROMPT;
-import static org.opentestsystem.ap.common.model.ItemConstants.ItemType.TYPE_MS;
+import static org.opentestsystem.ap.common.model.ModelConstants.ItemType.TYPE_MS;
 import static org.opentestsystem.ap.imrt.iis.core.keyword.KeywordContentUtil.convertContentToKeywordString;
 
 @Component
