+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/keyword/SaKeywordContentMapper.java
@@ -14,7 +14,7 @@ import java.util.Set;
 
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.SaSection.PROMPT;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.SaSection.SPANISH_PROMPT;
-import static org.opentestsystem.ap.common.model.ItemConstants.ItemType.TYPE_SA;
+import static org.opentestsystem.ap.common.model.ModelConstants.ItemType.TYPE_SA;
 import static org.opentestsystem.ap.imrt.iis.core.keyword.KeywordContentUtil.convertContentToKeywordString;
 
 @Component
