+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/keyword/StimKeywordContentMapper.java
@@ -14,7 +14,7 @@ import java.util.Set;
 
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.StimSection.CONTENT;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.StimSection.SPANISH_CONTENT;
-import static org.opentestsystem.ap.common.model.ItemConstants.ItemType.TYPE_STIM;
+import static org.opentestsystem.ap.common.model.ModelConstants.ItemType.TYPE_STIM;
 import static org.opentestsystem.ap.imrt.iis.core.keyword.KeywordContentUtil.convertContentToKeywordString;
 
 @Component
