+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/keyword/TiKeywordContentMapper.java
@@ -25,7 +25,7 @@ import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.TiSe
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.TiSection.TABLE_CELLS;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.TiSection.TABLE_COL_HEADERS;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.TiSection.TABLE_TITLE;
-import static org.opentestsystem.ap.common.model.ItemConstants.ItemType.TYPE_TI;
+import static org.opentestsystem.ap.common.model.ModelConstants.ItemType.TYPE_TI;
 import static org.opentestsystem.ap.imrt.iis.core.keyword.KeywordContentUtil.convertContentToKeywordString;
 
 @Component
