+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/keyword/TutKeywordContentMapper.java
@@ -14,7 +14,7 @@ import java.util.Set;
 
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.TutSection.CONTENT;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.TutSection.SPANISH_CONTENT;
-import static org.opentestsystem.ap.common.model.ItemConstants.ItemType.TYPE_TUT;
+import static org.opentestsystem.ap.common.model.ModelConstants.ItemType.TYPE_TUT;
 import static org.opentestsystem.ap.imrt.iis.core.keyword.KeywordContentUtil.convertContentToKeywordString;
 
 @Component
