+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/keyword/WerKeywordContentMapper.java
@@ -14,7 +14,7 @@ import java.util.Set;
 
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.WerSection.PROMPT;
 import static org.opentestsystem.ap.common.imrt.model.KeywordContentSection.WerSection.SPANISH_PROMPT;
-import static org.opentestsystem.ap.common.model.ItemConstants.ItemType.TYPE_WER;
+import static org.opentestsystem.ap.common.model.ModelConstants.ItemType.TYPE_WER;
 import static org.opentestsystem.ap.imrt.iis.core.keyword.KeywordContentUtil.convertContentToKeywordString;
 
 @Component
