+++ src/test/java/org/opentestsystem/ap/imrt/iis/core/keyword/MiKeywordContentMapperTest.java
@@ -3,7 +3,6 @@ package org.opentestsystem.ap.imrt.iis.core.keyword;
 import org.junit.Test;
 import org.opentestsystem.ap.common.imrt.model.BaseItem;
 import org.opentestsystem.ap.common.imrt.model.KeywordContent;
-import org.opentestsystem.ap.common.model.ItemConstants;
 import org.opentestsystem.ap.common.model.MiItem;
 import org.opentestsystem.ap.common.model.Table;
 import org.opentestsystem.ap.common.model.TiItem;
